@@ -31,11 +31,10 @@ pub fn kill(message: KillMessage, sender: &Sender<Message>, state: &SharedState)
                 &state,
             ),
             TaskSelection::Group(group) => create_success_message(format!(
-                "Sending signal {} to all running tasks of group {}.",
-                signal, group
+                "Sending signal {signal} to all running tasks of group {group}.",
             )),
             TaskSelection::All => {
-                create_success_message(format!("Sending signal {} to all running tasks.", signal))
+                create_success_message(format!("Sending signal {signal} to all running tasks."))
             }
         }
     } else {
@@ -46,10 +45,9 @@ pub fn kill(message: KillMessage, sender: &Sender<Message>, state: &SharedState)
                 |task| task.is_running(),
                 &state,
             ),
-            TaskSelection::Group(group) => create_success_message(format!(
-                "All tasks of group \"{}\" are being killed.",
-                group
-            )),
+            TaskSelection::Group(group) => {
+                create_success_message(format!("All tasks of group \"{group}\" are being killed."))
+            }
             TaskSelection::All => create_success_message("All tasks are being killed."),
         }
     }
