@@ -28,14 +28,14 @@ impl TaskHandler {
                 parallel_tasks,
             } => {
                 if state.groups.contains_key(&name) {
-                    error!("Group \"{}\" already exists", name);
+                    error!("Group \"{name}\" already exists");
                     return;
                 }
                 let mut group = state.create_group(&name);
                 if let Some(parallel_tasks) = parallel_tasks {
                     group.parallel_tasks = parallel_tasks;
                 }
-                info!("New group \"{}\" has been created", &name);
+                info!("New group \"{name}\" has been created");
 
                 // Create the worker pool.
                 self.children.0.insert(name, BTreeMap::new());
@@ -45,21 +45,18 @@ impl TaskHandler {
             }
             GroupMessage::Remove(group) => {
                 if !state.groups.contains_key(&group) {
-                    error!("Group \"{}\" to be remove doesn't exists", group);
+                    error!("Group \"{group}\" to be remove doesn't exists");
                     return;
                 }
 
                 // Make sure there are no tasks in that group.
                 if state.tasks.iter().any(|(_, task)| task.group == group) {
-                    error!(
-                        "Tried to remove group \"{}\", while it still contained tasks.",
-                        group
-                    );
+                    error!("Tried to remove group \"{group}\", while it still contained tasks.");
                     return;
                 }
 
                 if let Err(error) = state.remove_group(&group) {
-                    error!("Error while removing group: \"{}\"", error);
+                    error!("Error while removing group: \"{error}\"");
                     return;
                 }
 
@@ -84,7 +81,7 @@ impl TaskHandler {
                 // Persist the state.
                 ok_or_shutdown!(self, save_state(&state));
 
-                info!("Group \"{}\" has been removed", &group);
+                info!("Group \"{group}\" has been removed");
             }
         }
     }
